4.21*: bloglist expansion, step9

+++ part4/controllers/blogs.js
@@ -10,8 +10,22 @@ blogsRouter.get("/", async (request, response) => {
 
 blogsRouter.delete("/:id", async (request, response) => {
   const id = request.params.id;
-  const blogs = await Blog.findByIdAndRemove(id);
-  response.status(204).end();
+  const token = request.token;
+
+  const decodedToken = jwt.verify(token, process.env.SECRET);
+  if (!token || !decodedToken.id) {
+    return response.status(401).json({ error: "token missing or invalid" });
+  }
+  const user = await User.findById(decodedToken.id);
+  const blog = await Blog.findById(id);
+  if (blog.user.toString() === user.id.toString()) {
+    const blogs = await Blog.findByIdAndRemove(id);
+    response.status(204).end();
+  } else {
+    return response.status(403).json({
+      error: "You must be the creator of this blog in order to delete it",
+    });
+  }
 });
 
 blogsRouter.post("/", async (request, response) => {
